+++ Tests/XcodeTests/SwiftUIProject/SwiftUIProject/Previews.swift
@@ -1,0 +1,17 @@
+import SwiftUI
+
+struct PreviewOnlyView: View {
+    var body: some View {
+        Text("Hello, world!")
+    }
+}
+
+struct PreviewOnlyView_PreviewProvider: PreviewProvider {
+    static var previews: some View {
+        PreviewOnlyView()
+    }
+}
+
+#Preview {
+    PreviewOnlyView()
+}